+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - **BREAKING**: Database host maintenance feature.
 
+### Changed
+- **BREAKING**: Update database config.
+
 ## [0.25.0] - 2024-02-05
 
 ### Added